requirements_base.txt
# core dependencies
numpy==1.23.1
pandas==1.4.3

scipy==1.9.0
gmpy2==2.1.2

matplotlib==3.5.3

web3==5.30.0

requests==2.28.1

python-dotenv==0.21.0

# dev dependencies
black==22.6.0
flake8==5.0.4

pylint==2.14.5
astroid==2.11.7

scriv==0.16.0
bump2version==1.0.1

titanoboa
pytest